+++ src/components/list/list-item.test.tsx
@@ -8,6 +8,7 @@
 import * as React from 'react';
 
 import {createTestApp} from '../../../test-utils/create-test-app';
+import {Typography} from '../typography';
 
 import {ListItem} from './list-item';
 
@@ -19,22 +20,27 @@ afterEach(app.afterEach);
 describe('<ListItem />', () => {
     it('should render div HTML element', () => {
         const wrapper = app.mount(<ListItem />);
-        expect(wrapper.find('div')).toHaveLength(1);
+        expect(wrapper.find('li')).toHaveLength(1);
     });
 
-    it('should be focusable', () => {
+    it('should not be focusable by default', () => {
         const wrapper = app.mount(<ListItem />);
-        expect(wrapper.find('div').props().tabIndex).toBe(0);
+        expect(wrapper.find('li').props().tabIndex).toBe(undefined);
+    });
+
+    it('should be focusable when interactive', () => {
+        const wrapper = app.mount(<ListItem interactive />);
+        expect(wrapper.find('li').props().tabIndex).toBe(0);
     });
 
     it('should be not focusable if disabled', () => {
-        const wrapper = app.mount(<ListItem disabled />);
-        expect(wrapper.find('div').props().tabIndex).toBe(-1);
+        const wrapper = app.mount(<ListItem disabled interactive />);
+        expect(wrapper.find('li').props().tabIndex).toBe(-1);
     });
 
     it('should accept tabIndex', () => {
         const wrapper = app.mount(<ListItem tabIndex={2} />);
-        expect(wrapper.find('div').props().tabIndex).toBe(2);
+        expect(wrapper.find('li').props().tabIndex).toBe(2);
     });
 
     it('should render custom elements', () => {
@@ -46,5 +52,26 @@ describe('<ListItem />', () => {
         app.mount(<ListItem id="test" href="foo" />);
         expect(document.getElementById('test')).toBeInstanceOf(HTMLAnchorElement);
         expect(document.getElementById('test')).toHaveProperty('href', 'http://localhost/foo');
+    });
+
+    it('should render prepend element', () => {
+        app.mount(<ListItem prepend={<div id="prepend" />} />);
+        expect(document.getElementById('prepend')).toBeInstanceOf(HTMLElement);
+    });
+
+    it('should render append element', () => {
+        app.mount(<ListItem append={<div id="append" />} />);
+        expect(document.getElementById('append')).toBeInstanceOf(HTMLElement);
+    });
+
+    it('should render subtitle', () => {
+        const wrapper = app.mount(
+            <ListItem subtitle="subtitle" subtitleProps={{id: 'subtitle'}} />
+        );
+        const subtitle = wrapper.findWhere(
+            (element) => element.type() === Typography && element.props().id === 'subtitle'
+        );
+
+        expect(subtitle.props().children).toEqual('subtitle');
     });
 });